Refactor Live API Visitor Profile logic in own class
